requirements.txt
rembg[cpu]>=2.0.50
Pillow>=10.0.0
tqdm>=4.66.0
customtkinter>=5.2.0
